avm1: Moved Avm1 struct into its own file, fixing field access as needed
