+++ core/src/avm1/globals/as_broadcaster.rs
@@ -163,7 +163,7 @@ pub fn initialize<'gc>(
         initialize_internal(
             activation.context.gc_context,
             broadcaster,
-            activation.context.avm1.broadcaster_functions,
+            activation.context.avm1.broadcaster_functions(),
             activation.context.avm1.prototypes().array,
         );
     }
