+++ core/src/avm1/globals/context_menu.rs
@@ -27,7 +27,7 @@ pub fn constructor<'gc>(
 
     let built_in_items = ScriptObject::new(
         activation.context.gc_context,
-        Some(activation.context.avm1.prototypes.object),
+        Some(activation.context.avm1.prototypes().object),
     );
 
     built_in_items.set("print", true.into(), activation)?;
@@ -41,7 +41,7 @@ pub fn constructor<'gc>(
 
     this.set("builtInItems", built_in_items.into(), activation)?;
 
-    let constructor = activation.context.avm1.prototypes.array_constructor;
+    let constructor = activation.context.avm1.prototypes().array_constructor;
     let custom_items = constructor.construct(activation, &[])?;
 
     this.set("customItems", custom_items, activation)?;
@@ -58,7 +58,11 @@ pub fn copy<'gc>(
         .get("onSelect", activation)?
         .coerce_to_object(activation);
 
-    let constructor = activation.context.avm1.prototypes.context_menu_constructor;
+    let constructor = activation
+        .context
+        .avm1
+        .prototypes()
+        .context_menu_constructor;
     let copy = constructor
         .construct(activation, &[callback.into()])?
         .coerce_to_object(activation);
