+++ core/src/avm1/globals/displacement_map_filter.rs
@@ -190,7 +190,7 @@ pub fn map_point<'gc>(
     if let Some(object) = this.as_displacement_map_filter_object() {
         let (x, y) = object.map_point();
 
-        let proto = activation.context.avm1.prototypes.point_constructor;
+        let proto = activation.context.avm1.prototypes().point_constructor;
         let point = proto.construct(activation, &[x.into(), y.into()])?;
         return Ok(point);
     }
