+++ core/src/avm1/globals/function.rs
@@ -42,7 +42,7 @@ pub fn call<'gc>(
     myargs: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
     let this = match myargs.get(0).unwrap_or(&Value::Undefined) {
-        Value::Undefined | Value::Null => activation.context.avm1.globals,
+        Value::Undefined | Value::Null => activation.context.avm1.global_object_cell(),
         this_val => this_val.coerce_to_object(activation),
     };
     let empty = [];
@@ -73,7 +73,7 @@ pub fn apply<'gc>(
     myargs: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
     let this = match myargs.get(0).unwrap_or(&Value::Undefined) {
-        Value::Undefined | Value::Null => activation.context.avm1.globals,
+        Value::Undefined | Value::Null => activation.context.avm1.global_object_cell(),
         this_val => this_val.coerce_to_object(activation),
     };
     let args_object = myargs.get(1).cloned().unwrap_or(Value::Undefined);
