+++ core/src/avm1/globals/selection.rs
@@ -2,8 +2,8 @@ use crate::avm1::activation::Activation;
 use crate::avm1::error::Error;
 use crate::avm1::globals::as_broadcaster::BroadcasterFunctions;
 use crate::avm1::property_decl::{define_properties_on, Declaration};
-use crate::avm1::{Object, ScriptObject, TDisplayObject, TObject, Value};
-use crate::display_object::{EditText, TextSelection};
+use crate::avm1::{Object, ScriptObject, TObject, Value};
+use crate::display_object::{EditText, TDisplayObject, TextSelection};
 use gc_arena::MutationContext;
 
 const OBJECT_DECLS: &[Declaration] = declare_properties! {
