+++ core/src/avm1/object/script_object.rs
@@ -654,7 +654,7 @@ mod tests {
                 activation.context.gc_context,
                 Executable::Native(|_avm, _this, _args| Ok("Virtual!".into())),
                 None,
-                activation.context.avm1.prototypes.function,
+                activation.context.avm1.prototypes().function,
             );
 
             object.as_script_object().unwrap().add_property(
@@ -680,7 +680,7 @@ mod tests {
                 activation.context.gc_context,
                 Executable::Native(|_avm, _this, _args| Ok("Virtual!".into())),
                 None,
-                activation.context.avm1.prototypes.function,
+                activation.context.avm1.prototypes().function,
             );
 
             object.as_script_object().unwrap().add_property(
@@ -736,7 +736,7 @@ mod tests {
                 activation.context.gc_context,
                 Executable::Native(|_avm, _this, _args| Ok(Value::Null)),
                 None,
-                activation.context.avm1.prototypes.function,
+                activation.context.avm1.prototypes().function,
             );
 
             object.as_script_object().unwrap().define_value(
