+++ core/src/display_object/edit_text.rs
@@ -2,8 +2,9 @@
 
 use crate::avm1::activation::{Activation as Avm1Activation, ActivationIdentifier};
 use crate::avm1::function::ExecutionReason;
+use crate::avm1::runtime::Avm1;
 use crate::avm1::{
-    Avm1, Object as Avm1Object, StageObject as Avm1StageObject, TObject as Avm1TObject,
+    Object as Avm1Object, StageObject as Avm1StageObject, TObject as Avm1TObject,
     Value as Avm1Value,
 };
 use crate::avm2::{
